simplify Operation class construction (#726)
* simplify *Operation creation

* move Specification classes to separate file
